Add interface for User schema
