+++ models/Content.ts
@@ -1,11 +1,5 @@
 import { model, Schema } from 'mongoose';
 
-enum userRole {
-    Admin = 0,
-    Writer = 1,
-    Reader = 2
-};
-
 enum contentType {
     Text = 0,
     Image = 1,
@@ -17,33 +11,6 @@ enum contentColumn {
     Middle = 1,
     Right = 2
 };
-
-const userSchema = new Schema({
-    userID: {
-        type: Number,
-        required: true
-    },
-    name: {
-        type: String,
-        required: true
-    },
-    email: {
-        type: String,
-        required: true
-    },
-    profilePicture: {
-        type: String,
-        required: true
-    },
-    role: {
-        type: userRole,
-        default: userRole.Reader
-    },
-    submissions: {
-        type: Array<Schema.Types.ObjectId>,
-        ref: "Content"
-    }
-});
 
 const contentSchema = new Schema({
     userID: {
@@ -92,5 +59,4 @@ const contentSchema = new Schema({
     }
 });
 
-export const User = model("User", userSchema);
 export const Content = model("Content", contentSchema);
